refactor(#1718): split tui/model.rs inline tests into model_tests.rs
model.rs was 1457 lines (limit 800) and this issue's +14 lines tripped the
file-size ratchet. Move the 781-line inline `mod tests` into
cqlite-cli/src/tui/model_tests.rs via the repo's #[path] convention:
model.rs 1457 -> 693, model_tests.rs 783 (test limit 1500). No test bodies
changed; all 50 tui unit tests still pass.

Campsite rule, epic #1116 / #1135.
